+++ docs/changelog.rst
@@ -35,7 +35,7 @@ Changed
 - Split ``WflowModel`` into ``WflowBaseModel`` containing all generic parts of all wflow models, like components and some setup functions, and ``WflowSbmModel``, containing the specific parts for the wflow_sbm concept.
 - ``WflowSedimentModel`` and ``WflowSbmModel`` both inherit from ``WflowBaseModel`` and extend it for their specific use cases.
 - The old ``WflowModel.setup_rivers`` has been split into three parts: ``WflowBaseModel.setup_rivers``, ``WflowSbmModel.setup_rivers`` and ``WflowSbmModel.setup_river_roughness``. Where the first contains the generic river setup logic, the second Sbm-specific implementations, and the third contains manning roughness implementations.
-
+- Clipping a model is now done by calling a single **clip** method. This function can be called with the ``hydromt update`` cli command.
 
 
 Unreleased